Dev: added navigation and routing to app and finished learing basic concepts of Vuejs
