+++ src/pages/HomePage.vue
@@ -1,0 +1,9 @@
+<template>
+    <h1>Главная страница</h1>
+</template>
+
+<script>
+export default {}
+</script>
+
+<style></style>
