+++ src/pages/PostIdPage.vue
@@ -1,0 +1,11 @@
+<template>
+    <div>
+        <h1>Страница поста с ID = {{ $route.params.id }}</h1>
+    </div>
+</template>
+
+<script>
+export default {}
+</script>
+
+<style></style>
