+++ src/router/routes.js
@@ -1,0 +1,3 @@
+export const HOME_PAGE_URL = '/'
+export const POSTS_PAGE_URL = '/posts'
+export const POSTS_ID_PAGE_URL = '/posts/:id'
